% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeansCpp}
\alias{kmeansCpp}
\title{K-Means clustering using Rcpp}
\usage{
kmeansCpp(data, k, maxIter = 100L)
}
\arguments{
\item{data}{numeric matrix of data, each row representing a sample}

\item{k}{the number of clusters}

\item{maxIter}{the maximum number of iterations allowed}
}
\value{
an integer cluster vector, with each integer representing a group.
}
\description{
Perform k-means clustering on a data matrix using Rcpp.
}
\examples{
dat <- as.matrix(iris[,1:2])
kmeansCpp(dat, k=3)
}
