% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansR.R
\name{kmeansR}
\alias{kmeansR}
\title{K-Means clustering using R}
\usage{
kmeansR(data, k, maxIter = 100)
}
\arguments{
\item{data}{numeric matrix of data, each row representing a sample}

\item{k}{the number of clusters}

\item{maxIter}{the maximum number of iterations allowed}
}
\value{
an integer cluster vector, with each integer representing a group.
}
\description{
Perform k-means clustering on a data matrix using R.
}
\examples{
dat <- as.matrix(iris[,1:2])
kmeansR(dat, k=3)
}
