\name{SA23204169-package}
\alias{SA23204169-package}
\alias{SA23204169}
\docType{package}
\title{The final project for 'Statistical Computing'}
\description{Summarise all homework in this course, implement K-means with R and Rcpp by myself and compare their efficiency with built-in function kmeans, implement the subgroup method in a 2016 paper.}
\section{Package Content}{
\tabular{ll}{
\strong{Title} \tab \strong{Description} \cr
\link[=SA23204169-package]{SA23204169-package} \tab The final project for 'Statistical Computing' \cr
\link[=SubgroupSolver-class]{SubgroupSolver-class} \tab A subgroup solver based on ADMM \cr
\link[=benchmarks]{benchmarks} \tab Benchmark R and Rcpp functions \cr
\link[=kmeansCpp]{kmeansCpp} \tab K-Means clustering using Rcpp \cr
\link[=kmeansR]{kmeansR} \tab K-Means clustering using R
}
}
\section{Vignettes}{
\tabular{ll}{
\strong{Title} \tab \strong{Description} \cr
\href{../doc/Homework.html}{Homework} \tab Homework of Statistical Computing. \cr
\href{../doc/PackageExample.html}{PackageExample} \tab Provide some user examples for the functions in this package.
}
}
\author{
Huiyang Peng
}
\keyword{package}
