% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subgroup.R
\docType{class}
\name{SubgroupSolver-class}
\alias{SubgroupSolver-class}
\alias{SubgroupSolver}
\title{A subgroup solver based on ADMM}
\description{
Here I implement the method proposed by Shujie Ma & Jian Huang in 2016.
}
\section{Methods}{

\describe{
\item{\code{fit(method, lamb, gamma, theta, eps = 0.001, omega = NULL)}}{Giving the solver the method and some parameters and start to solve the subgroup problem.

 The parameter \code{method} must be one of ('MCP','SCAD' and 'L1'), which applies to different penalties in the paper.
 
 MCP:
 
 \deqn{p_{\gamma}(t,\lambda)=\lambda\int_0^t(1-x/(\gamma\lambda))_+ dx}
 
 SCAD:
 
 \deqn{p_{\gamma}(t,\lambda)=\lambda\int_0^t\min\{1,(\gamma-x/\lambda)_+ /(\gamma-1)\} dx}
 
 L1:
 
 \deqn{p_{\omega}(t,\lambda)=\lambda \omega_{ij} |t|}
 
 and \code{theta} is another parameter in the process of ADMM algorithm.}

\item{\code{initialize(X, y)}}{Initialize a SubgroupSolver.

This function creates and initializes a new instance of the SubgroupSolver class. 

Here X is a numeric matrix and y is a numeric vector, we apply \code{solver = SubgroupSolver(X,y)} to initialize a object.}

\item{\code{update(method, lamb, gamma, theta, omega = NULL)}}{Update the solver under some given parameters, usually useless for user.}
}}

\examples{
set.seed(1)
covmatrix <- matrix(0.3, ncol = 5, nrow = 5)
diag(covmatrix) <- 1
meanvalue <- rep(0,5)
X <- mvtnorm::rmvnorm(100, meanvalue, covmatrix)
beta <- runif(5, min = 0.5, max = 1)
mu <- sample(c(-2, 2), size = 100, replace = TRUE)
eps <- rnorm(100, 0, 0.5)
y <- as.vector(X \%*\% beta) + mu + eps
solver <- SubgroupSolver(X,y)
solver$fit("MCP",0.6, 3, 1)
solver$mu
}
\references{
Shujie Ma & Jian Huang(2016).A concave pairwise fusion approach to subgroup analysis
}
