\name{SA23204169-package}
\alias{SA23204169-package}
\alias{SA23204169}
\docType{package}
\title{The final project for 'Statistical Computing'}
\description{Summarise all homework in this course, implement K-means with R and Rcpp by myself and compare their efficiency with built-in function kmeans, implement the subgroup method in a 2016 paper.}
\section{Package Content}{
\tabular{ll}{
\strong{Title} \tab \strong{Description} \cr
\link[=SA23204169-package]{SA23204169-package} \tab The final project for 'Statistical Computing' \cr
\link[=SubgroupSolver-class]{SubgroupSolver-class} \tab A subgroup solver based on ADMM \cr
\link[=benchmarks]{benchmarks} \tab Benchmark R and Rcpp functions \cr
\link[=kmeansCpp]{kmeansCpp} \tab K-Means clustering using Rcpp \cr
\link[=kmeansR]{kmeansR} \tab K-Means clustering using R
}
}
\section{Vignettes}{
\tabular{ll}{
\strong{Title} \tab \strong{Description} \cr
\href{../doc/A-SA23204169-2023-9-11.html}{A-SA23204169-2023-9-11} \tab Homework-2023-9-11 \cr
\href{../doc/A-SA23204169-2023-9-18.html}{A-SA23204169-2023-9-18} \tab Homework-2023-9-18 \cr
\href{../doc/A-SA23204169-2023-9-25.html}{A-SA23204169-2023-9-25} \tab Homework-2023-9-25 \cr
\href{../doc/A-SA23204169-2023-10-9.html}{A-SA23204169-2023-10-9} \tab Homework-2023-10-9 \cr
\href{../doc/A-SA23204169-2023-10-16.html}{A-SA23204169-2023-10-16} \tab Homework-2023-10-16 \cr
\href{../doc/A-SA23204169-2023-10-23.html}{A-SA23204169-2023-10-23} \tab Homework-2023-10-23 \cr
\href{../doc/A-SA23204169-2023-10-30.html}{A-SA23204169-2023-10-30} \tab Homework-2023-10-30 \cr
\href{../doc/A-SA23204169-2023-11-6.html}{A-SA23204169-2023-11-6} \tab Homework-2023-11-6 \cr
\href{../doc/A-SA23204169-2023-11-13.html}{A-SA23204169-2023-11-13} \tab Homework-2023-11-13 \cr
\href{../doc/A-SA23204169-2023-11-20.html}{A-SA23204169-2023-11-20} \tab Homework-2023-11-20 
}
}
\author{
Huiyang Peng
}
\keyword{package}
